# api-contract.yaml
openapi: 3.0.3
info:
  title: Debt Management System API
  description: >
    Автоматизированная система управления дебиторской задолженностью
    с модулем омниканальных уведомлений (SMS / WhatsApp / Telegram / Email).
  version: 1.0.0
  contact:
    name: Dev Team
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
  - url: https://api.debtmanager.example.com/v1
    description: Production
...
paths:
...
  /roles/:
    get:
      tags: [Roles]
      summary: Список ролей
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Role"
    post:
      tags: [Roles]
      summary: Создать роль
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreate"
      responses:
        "201":
          description: Роль создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
  /roles/{id}/permissions/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    put:
      tags: [Roles]
      summary: Назначить права роли (полная замена)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permission_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
      responses:
        "200":
          description: Права обновлены
...
  /notification-templates/:
    get:
      tags: [Notifications]
      summary: Список шаблонов
      security:
        - BearerAuth: []
      parameters:
        - name: channel
          in: query
          schema:
            type: string
            enum: [whatsapp, sms, telegram, email]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationTemplate"
    post:
      tags: [Notifications]
      summary: Создать шаблон
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationTemplateCreate"
      responses:
        "201":
          description: Создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationTemplate"
  /notification-templates/{id}/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    get:
      tags: [Notifications]
      summary: Получить шаблон
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationTemplate"
    patch:
      tags: [Notifications]
      summary: Обновить шаблон
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationTemplateCreate"
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationTemplate"
    delete:
      tags: [Notifications]
      summary: Удалить шаблон
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Удалено
  /notifications/send/:
    post:
      tags: [Notifications]
      summary: Ручная отправка уведомления
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [debt_case_id, template_id, channel]
              properties:
                debt_case_id:
                  type: string
                  format: uuid
                template_id:
                  type: string
                  format: uuid
                channel:
                  type: string
                  enum: [whatsapp, sms, telegram, email]
      responses:
        "202":
          description: Уведомление поставлено в очередь
          content:
            application/json:
              schema:
                type: object
                properties:
                  log_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: queued
  /notifications/logs/:
    get:
      tags: [Notifications]
      summary: Лог отправленных уведомлений
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: debt_case_id
          in: query
          schema:
            type: string
            format: uuid
        - name: channel
          in: query
          schema:
            type: string
            enum: [whatsapp, sms, telegram, email]
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, sent, delivered, failed]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedNotificationLogs"

  # ── SCHEDULER ─────────────────────────
  /scheduled-tasks/:
    get:
      tags: [Scheduler]
      summary: Список запланированных задач
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: task_status
          in: query
          schema:
            type: string
            enum: [pending, sent, cancelled, failed]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedScheduledTasks"
    post:
      tags: [Scheduler]
      summary: Создать задачу рассылки
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledTaskCreate"
      responses:
        "201":
          description: Задача создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledTask"
  /scheduled-tasks/{id}/cancel/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    post:
      tags: [Scheduler]
      summary: Отменить задачу
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Задача отменена
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: cancelled

  # ── PTP ───────────────────────────────
  /ptp/:
    get:
      tags: [PTP]
      summary: Список обещаний об оплате
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, kept, broken]
        - name: debt_case_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedPTPRecords"
    post:
      tags: [PTP]
      summary: Зафиксировать обещание об оплате
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PTPCreate"
      responses:
        "201":
          description: Создано
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTPRecord"
  /ptp/{id}/status/:
    parameters:
      - $ref: "#/components/parameters/UUIDParam"
    patch:
      tags: [PTP]
      summary: Обновить статус PTP
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [kept, broken]
                note:
                  type: string
      responses:
        "200":
          description: Статус обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PTPRecord"

  # ── REPORTS ───────────────────────────
  /reports/campaign/:
    get:
      tags: [Reports]
      summary: Отчёт по кампаниям рассылки
      security:
        - BearerAuth: []
      parameters:
        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2025-05-01"
        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2025-05-31"
        - name: channel
          in: query
          schema:
            type: string
            enum: [whatsapp, sms, telegram, email, all]
            default: all
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignReport"
  /reports/agent-activity/:
    get:
      tags: [Reports]
      summary: Лог активности агентов
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - name: agent_id
          in: query
          schema:
            type: string
            format: uuid
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedAgentActivityLogs"
  /reports/summary/:
    get:
      tags: [Reports]
      summary: Сводная статистика (dashboard)
      security:
        - BearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  # ── INTEGRATIONS ──────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UUIDParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        created_at:
          type: string
          format: date-time
...
    Role:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: agent
        description:
          type: string
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
    RoleCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    Permission:
      type: object
      properties:
        id:
          type: string
          format: uuid
        codename:
          type: string
          example: can_send_notification
        description:
          type: string

    # ── DEBTOR ──
...
    NotificationTemplate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Напоминание DPD 1-7 дней"
        channel:
          type: string
          enum: [whatsapp, sms, telegram, email]
        body:
          type: string
          example: "Уважаемый {{full_name}}, у вас задолженность {{amount}} сом. Пожалуйста, оплатите до {{due_date}}."
        language:
          type: string
          example: ru
    NotificationTemplateCreate:
      type: object
      required: [name, channel, body]
      properties:
        name:
          type: string
        channel:
          type: string
          enum: [whatsapp, sms, telegram, email]
        body:
          type: string
        language:
          type: string
          default: ru
    NotificationLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        debt_case_id:
          type: string
          format: uuid
        template:
          $ref: "#/components/schemas/NotificationTemplate"
        channel:
          type: string
        status:
          type: string
          enum: [queued, sent, delivered, failed]
        sent_at:
          type: string
          format: date-time
          nullable: true
        response_raw:
          type: string
          nullable: true
          description: Сырой ответ от провайдера
    PaginatedNotificationLogs:
      allOf:
        - $ref: "#/components/schemas/Pagination"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/NotificationLog"

    # ── SCHEDULER ──
    ScheduledTask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        debt_case_id:
          type: string
          format: uuid
        template:
          $ref: "#/components/schemas/NotificationTemplate"
        channel:
          type: string
          enum: [whatsapp, sms, telegram, email]
        scheduled_at:
          type: string
          format: date-time
        task_status:
          type: string
          enum: [pending, sent, cancelled, failed]
    ScheduledTaskCreate:
      type: object
      required: [debt_case_id, template_id, channel, scheduled_at]
      properties:
        debt_case_id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        channel:
          type: string
          enum: [whatsapp, sms, telegram, email]
        scheduled_at:
          type: string
          format: date-time
    PaginatedScheduledTasks:
      allOf:
        - $ref: "#/components/schemas/Pagination"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/ScheduledTask"

    # ── PTP ──
    PTPRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        debt_case_id:
          type: string
          format: uuid
        agent:
          $ref: "#/components/schemas/User"
        promise_date:
          type: string
          format: date
        promised_amount:
          type: number
          format: float
        status:
          type: string
          enum: [pending, kept, broken]
        created_at:
          type: string
          format: date-time
    PTPCreate:
      type: object
      required: [debt_case_id, promise_date, promised_amount]
      properties:
        debt_case_id:
          type: string
          format: uuid
        promise_date:
          type: string
          format: date
        promised_amount:
          type: number
          format: float
    PaginatedPTPRecords:
      allOf:
        - $ref: "#/components/schemas/Pagination"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/PTPRecord"

    # ── REPORTS ──
    CampaignReport:
      type: object
      properties:
        report_date:
          type: string
          format: date
        channel:
          type: string
        total_sent:
          type: integer
        total_delivered:
          type: integer
        total_ptp:
          type: integer
        collected_amount:
          type: number
          format: float
    AgentActivityLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        agent:
          $ref: "#/components/schemas/User"
        debt_case_id:
          type: string
          format: uuid
        action_type:
          type: string
          example: manual_notification_sent
        performed_at:
          type: string
          format: date-time
        note:
          type: string
          nullable: true
    PaginatedAgentActivityLogs:
      allOf:
        - $ref: "#/components/schemas/Pagination"
        - type: object
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/AgentActivityLog"
    DashboardSummary:
      type: object
      properties:
        total_open_cases:
          type: integer
        total_overdue_cases:
          type: integer
        total_amount_outstanding:
          type: number
          format: float
        ptp_pending:
          type: integer
        ptp_kept_this_month:
          type: integer
        notifications_sent_today:
          type: integer
        collection_rate_percent:
          type: number
          format: float
          example: 34.5

    # ── INTEGRATIONS ──
...
    Pagination:
      type: object
      properties:
        count:
          type: integer
          description: Всего объектов
        next:
          type: string
          format: uri
          nullable: true
        previous:
          type: string
          format: uri
          nullable: true